addorder put operation complete

+++ src/app/modules/user/user.service.ts
@@ -27,6 +27,9 @@ const updateUser = async (
   userId: number,
   userData: Iuser,
 ): Promise<Iuser | null> => {
+  if ((await User.isUserExists(userId)) === null) {
+    throw new Error('User donot exists!');
+  }
   const result = await User.findOneAndUpdate({ userId }, userData, {
     new: true,
   });
@@ -35,13 +38,32 @@ const updateUser = async (
 };
 
 const deleteUser = async (userId: number): Promise<Iuser | null> => {
+  if ((await User.isUserExists(userId)) === null) {
+    throw new Error('User donot exists!');
+  }
   const result = await User.findOneAndDelete({ userId });
   return result;
 };
+
+const addOrder = async (userId: number, orderData) => {
+  if ((await User.isUserExists(userId)) === null) {
+    throw new Error('User donot exists!');
+  }
+
+  const result = await User.findOneAndUpdate(
+    { userId },
+    { $push: { orders: orderData } },
+    { new: true, select: 'orders' },
+  );
+
+  return result;
+};
+
 export const UserManagementService = {
   createUser,
   getAllUsers,
   specificUserById,
   updateUser,
   deleteUser,
+  addOrder,
 };
